Disable func-names for ES5/ES6
This rule was able to be partially avoided in ES6 by using arrow
functions `() =>`, but using generator functions `function*()` requires
a function signature. This rule is not very helpful to most of our use
cases, since we use anonymous functions extensively for callbacks.
Therefore, we are disabling it for ES5/ES6.