Enable object-shorthand (#2)
For some reason, the team decided to disable object-shorthand when we
developed our styleguide. Since we have spent most of our time the past
couple months in ES5 on Console, working more with API and default
configuration showed that this is a useful feature that should stay
defaulted to Airbnb's style.